fix: build on macOS with Apple clang
- server: use eventfd on Linux and a self-pipe elsewhere (eventfd is
  Linux-only) to wake the select loop from the signal handler
- utils: drop glibc-only <bits/types/struct_timeval.h> and use standard
  isinf instead of the GNU isinff
- anet: drop glibc-only <asm-generic/...> includes
- core: print u64 dims with %llu to match its type on all platforms

+++ src/core.c
@@ -1807,7 +1807,7 @@ static void model_summary(Model *m) {
         fprintf(stdout, "\t\t= (%s)", gguf_types[t->type].name);
         fprintf(stdout, "[");
         for (u32 j = 0; j < t->ndim; j++) {
-            fprintf(stdout, "%ld", t->dim[j]);
+            fprintf(stdout, "%llu", (unsigned long long)t->dim[j]);
             if (j < t->ndim - 1) fprintf(stdout, ",");
         }
         fprintf(stdout, "]\n");
